move profile state to state directory
